HomeWork/hw7/requirements.txt
Django==3.1
djangorestframework==3.12.4
requests==2.31.0
djangorestframework-authtoken==1.1.0
python-dotenv==1.0.1
